requirements.txt
pydantic-settings==2.9.1
dash-bootstrap-components==2.0.2
Flask-Session==0.8.0
celery==5.5.2
redis==6.1.0
SQLAlchemy==2.0.41
